Adjusting sessionDb and userDb IDs to be 64-bit in the code, since the database is now ready for it and we're crossing the threshold.  refs #33554

+++ src/hg/hgTrackUi/hgTrackUi.c
@@ -3043,7 +3043,7 @@ freeMem(tdbDataTypes);
 return list;
 }
 
-unsigned int cartDbParseId(char *, char **);  // ADS: avoid extra include
+unsigned long cartDbParseId(char *, char **);  // ADS: avoid extra include
 
 
 static void facetedCompositeUi(struct trackDb *tdb)
